image delete file and store in volume

+++ src/images/images.service.ts
@@ -4,10 +4,14 @@ import { ImagesRepository } from "./images.repository";
 import { Image } from "./schemas/image.schema";
 import { CreateImageDto } from "./dto/create-image.dto";
 import { UpdateImageDto } from "./dto/update-image.dto";
+import { FileStorageService } from "@/file-storage/file-storage.service";
 
 @Injectable()
 export class ImagesService {
-  constructor(private readonly imagesRepository: ImagesRepository) {}
+  constructor(
+    private readonly imagesRepository: ImagesRepository,
+    private readonly fileStorageService: FileStorageService
+  ) {}
 
   create(createImageDto: CreateImageDto): Promise<Image> {
     return this.imagesRepository.create(createImageDto);
@@ -25,7 +29,15 @@ export class ImagesService {
     return this.imagesRepository.updateById(id, updateTemplateDto);
   }
 
-  remove(id: string): Promise<Image | null> {
-    return this.imagesRepository.remove(id);
+  async remove(id: string): Promise<Image | null> {
+    const image = await this.imagesRepository.findById(id);
+
+    if (!image) {
+      return null;
+    }
+
+    this.fileStorageService.deleteFile(image?.filename);
+
+    return await this.imagesRepository.remove(id);
   }
 }
